requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
kafka-python==2.0.2
requests==2.32.3
six==1.16.0
urllib3==2.2.1
